develop

+++ backend/SkillProof/SkillProof.Entities/Dtos/Tests/QuestionResultDto.cs
@@ -8,7 +8,7 @@ public class QuestionResultDto
     public string QuestionTitle { get; set; } = string.Empty;
     public QuestionType Type { get; set; }
     public bool IsCorrect { get; set; }
-    public int PointsAwarded { get; set; }
+    public double PointsAwarded { get; set; }
     public int MaxPoints { get; set; }
     public string UserResponse { get; set; } = string.Empty;
     public string AiFeedback { get; set; } = string.Empty;
